[REFACTOR] #69 - Make SwipeTouchHelper
