+++ presentation/src/main/java/com/woowahan/repositorysearch/ui/main/fragment/NotificationFragment.kt
@@ -19,6 +19,7 @@ import com.woowahan.repositorysearch.R
 import com.woowahan.repositorysearch.databinding.FragmentNotificationBinding
 import com.woowahan.repositorysearch.ui.adapter.NotificationAdapter
 import com.woowahan.repositorysearch.ui.main.DividerItemDecoration
+import com.woowahan.repositorysearch.ui.recyclerview.SwipeTouchHelper
 import com.woowahan.repositorysearch.util.Dp2Px
 import dagger.hilt.android.AndroidEntryPoint
 import java.util.*
@@ -71,71 +72,18 @@ class NotificationFragment : Fragment() {
             val throwable = it[2] as Throwable
 
             notificationAdapter.restoreItem(notification, position)
-            Toast.makeText(requireContext(), throwable.message.toString(), Toast.LENGTH_SHORT).show()
+            Toast.makeText(requireContext(), throwable.message.toString(), Toast.LENGTH_SHORT)
+                .show()
         }
 
-        ItemTouchHelper(recyclerViewTouchCallback).attachToRecyclerView(binding.rvNotification)
+        ItemTouchHelper(SwipeTouchHelper {
+            viewModel.markNotificationAsRead(
+                it,
+                notificationAdapter.currentList[it]
+            )
+            notificationAdapter.removeItem(it)
+        }).attachToRecyclerView(binding.rvNotification)
     }
 
-    private val recyclerViewTouchCallback =
-        object : ItemTouchHelper.SimpleCallback(0, ItemTouchHelper.LEFT) {
-            override fun onMove(
-                recyclerView: RecyclerView,
-                viewHolder: RecyclerView.ViewHolder,
-                target: RecyclerView.ViewHolder
-            ): Boolean {
-                return false
-            }
 
-            override fun onSwiped(viewHolder: RecyclerView.ViewHolder, direction: Int) {
-                val position = viewHolder.adapterPosition
-                viewModel.markNotificationAsRead(
-                    position,
-                    notificationAdapter.currentList[position]
-                )
-                notificationAdapter.removeItem(position)
-            }
-
-            override fun onChildDraw(
-                c: Canvas,
-                recyclerView: RecyclerView,
-                viewHolder: RecyclerView.ViewHolder,
-                dX: Float,
-                dY: Float,
-                actionState: Int,
-                isActiveNow: Boolean
-            ) {
-                super.onChildDraw(c, recyclerView, viewHolder, dX, dY, actionState, isActiveNow)
-
-                if (actionState == ItemTouchHelper.ACTION_STATE_SWIPE) {
-                    val icon =
-                        ContextCompat.getDrawable(requireContext(), R.drawable.ic_check) ?: return
-
-                    val view = viewHolder.itemView
-                    val paint = Paint()
-                    paint.color = ContextCompat.getColor(requireContext(), R.color.primary)
-
-                    val icPadding = Dp2Px.convert(requireContext(), 48F)
-                    val icLeft = (view.right - icPadding - icon.intrinsicWidth).toInt()
-                    val icTop = view.top + ((view.bottom - view.top - icon.intrinsicHeight) / 2)
-                    val icRight = (view.right - icPadding).toInt()
-                    val icBottom = icTop + icon.intrinsicHeight
-
-                    c.drawRect(
-                        view.right.toFloat() + dX,
-                        view.top.toFloat(),
-                        view.right.toFloat(),
-                        view.bottom.toFloat(),
-                        paint
-                    )
-
-                    if (-dX < icon.intrinsicWidth + icPadding) {
-                        icon.bounds = Rect(0, 0, 0, 0)
-                    } else {
-                        icon.bounds = Rect(icLeft, icTop, icRight, icBottom)
-                        icon.draw(c)
-                    }
-                }
-            }
-        }
 }
